Feature: Change python dcl methods to produce ipython output
The dcl based methods in the python module currently pass NULL pointers for
output and error if they are not asked to return the output. This causes
the mdsdcl code to print the output to stdout and stderr directly. This PR
changes the code to obtain the output and error strings and use python's
print function to print the output string and the error string. This enables
things like ipython notebooks to display the output and errors in the notebook.