Build: Fix promote.sh script
deploy/promote.sh script is used to create a new promotion branch
to create a pull request to essentially rebase one development branch
onto another (i.e. rebase alpha on top of stable). The script now
performs the required rebasing and merging and the end result should be
a branch with the same head as the promote to branch while retaining the
history of commits from both branches. The script needs to be copied
to /tmp and executed there to prevent the script itself being replaced
in the midst of the rebasing and merging.