Fix: Repair rpm installation scripts
When splitting the debian and redhat installer scripts the build
script was inadvertently using the debian scripts for noarch packages
which led configuration changes to be removed during package updates.
On redhat the uninstall scripts of the installed package is invoked
after the updated package is installed and the redhat uninstall scripts
check an argument to determine if it is an update or a perminant package
deinstall so that it won't remove things if it is an update. The debian
packages don't have this problem. It will take a couple updates for the
fix to take effect though since the deinstall scripts of the installed
package will be invoked until the fixed scripts are added with new versions.