Fix: enable data(build_slope(...)) in tdi
Although the DTYPE_SLOPE data type has been deprecated for years they can
still be encountered in tree datafiles. It was discovered that attempting
to convert a DTYPE_SLOPE into an array of values by doing operations
such as: data(build_slope(.1,0.,1.))
caused a segfault in tditest. This fix should prevent this segfault
and return an array of values if the DTYPE_SLOPE contains values for the
increment, begin and end arguments for all of the slope segments.