Fix: Truncate last segment if not yet filled
If the last segment of a node containing regular (not timestamped)
segments is only partially filled a call to getSegment for the last
segment returned the contents of the whole segment instead of only
the filled portion. This PR should truncate the data and dimension
to return only the filled portion of the segment.