Fix: Require consistent dispatch table context
The dispatch table will be only valid when the context is referring
to the tree the dispatch/build command used to generate the table.
The dispatch table is now attached to the tree context so if you
open another tree or close the original tree you will get an error
if you try to issue the dispatch/phase. If you open another tree
and then set the tree context to the original the dispatch table
will be available.