Feature: Provide tdi based treeshr hooks
This is the initial addition of tdi based treeshr hooks and provides
hooks for all the existing hooks. New hooks will be added for segment
operations and possibly the addition of extended hooks which pass
more information to the hook such as the data being written to the tree.

The tdi based hooks will call a TreeShrHook function which receives a
Dictionary of hook related fields such as the hook type, tree, shot, nid,
etc. The hooks called will be selectable with an environment variable called
"TreeHooks". This can be a comma delimited list of hook types or "all" to
enable all hooks. If TreeHooks is not defined no hooks will be called.
This will reduce the overhead if no TreeShrHook function is defined and
if only a subset of the hook types are of interest.