Feature: Add tdi TreeHooks feature
Add various treeshr hooks which will call a TreeShrHook fun with
one argument which is a Dictionary. Depending on the hook the
Dictionary could contain fields such as:

  "type" - the name of the type of hook
  "tree" - the name of the tree
  "shot" - the shot number of the tree
  "nid"  - the node being operated on
  "data" - the data being written to the tree

The type of hooks currently provided are:

  "OpenTree"     - Tree being opened
  "WriteTree"    - Save of edited tree
  "GetNci"       - Getting node characteristics
  "GetData"      - Retrieving data stored in a node
  "CloseTree"    - Closing an opened tree
  "RetrieveTree" - Tree not found so hook to retrieve from storage
  "OpenTreeEdit" - Opening a tree for edit
  "OpenDatafileWrite" - Open tree datafile for write
  "MakeSegment"  - Store a segment in a node
  "MakeSegmentFull" - Store a segment in a node (data passed to hook)
  "MakeTimestampedSegment" - Store timestamped segment
  "MakeTimestampedSegmentFull" - Store segment (data passed)
  "PutSegment"      - Store segment
  "PutTimestampedSegment" - Store timestamped segment
  "PutTimestampedSegmentFull" - (data passed)

Hook selection is determined by the TreeHooks environment variable.
You can select particular hooks by listing the hook types in a
comma delimited list (i.e. export TreeHooks=OpenTree,GetNci,CloseTree)
or you can use the keywords allhooks and fullhooks. If the string,
"allhooks", is found in the TreeHooks environment variable all of the
hooks which do not have full in their type name will be called. If
the string, "fullhooks", is found all the hooks with full in their type
will be called. If you include allhooks and fullhooks in the TreeHooks
environment variable all the hooks will be called. The names included
in the TreeHooks environment variable will be case insensitive.