Feature: enable importing of other modules in MDS_PYDEVICE_PATH
The Device.importPyDeviceModule(name) class method is used to
import pydevice modules found in the MDS_PYDEVICE_PATH environment
variable. This originally only provided support for importing
name.py files. This adds the capability of importing packages
contained in the directory with that name.

This class method is case insensitive when searching for modules
with the specified name.

When a device implementation imports a module inside of a method
or imports a module which resides in a directory other than
the device module itself but inside the MDS_PYDEVICE_PATH
directory trees it should use the importPyDeviceModule() method
to import a module.

Once imported these modules are cached so future calls to
importPyDeviceModule() do not need to search the directory trees.
Similarly if a module is found but fails to import this failure
is also cached so subsequent attempts do not search for the module.