Feature: py3 install and verify python and library compatibility
This change will add the installation of the MDSplus python
module in python3 if it is available on redhat and debian platforms.
It also adds a version check when importing the MDSplus python
module into python. The MDSplus python modue will now check
the version of the installed libraries using the MDSplusVersion
structure now exported by the MdsShr library and compare that
with the version information stored in the _version.py module
included with the MDSplus python installer. If a user is running
the MDSplus python module in a virtual environment and the
MDSplus libraries are updated the user will get a warning when
importing the MDSplus module. It will be up to the user to
update the python module from the mdsplus/mdsobjects/python directory
of the installation directory.