Feature: enable python module library mismatch reporting
When importing the MDSplus python module it will now verify
that the module matches the version of the MDSplus libraries.
If the sources of the python module is not the same version
as the MDSplus libraries there could easily be compatibility
issues. The module will now warn if a mismatch is detected.