Feature: Add MDSplus-previous repo to enable downgrade
Even though the deb files for previous versions are in the debian
repositories they cannot be installed using apt-get to downgrade
to the previous release. This enables downgrades using

apt-get install mdsplus-xxxx=version

mdsplus-xxxx is the package you want to downgrade and version
is the version you want to downgrade to (i.e. 7.60.2)

This supports only downgrading to the previous version. To downgrade
to ealier versions you must download deb files from
http://www.mdsplus.org/dist directory trees.