Feature: Add "code in" pydevice support
This enables the feature to develop pydevices in which the device
methods are stored in the tree. The only time the python module file
is referenced is to add the device to an MDSplus tree. From that point
on the methods executed on the device are implemented in code living
in the trees datafile.