Build: Avoid need to copy python to build dir
The Makefiles were modified to copy all the python sources
into the build directory so the version checking of the python
modules would not break the tests if _version.py was not in the
python source directory tree. This causes problems with debugging
and testing changes to the python code in the source directory.
Instead this change adds an environment variable MDS_DIS_VCHECK
which can be used to disable the warning message done by the
version checking code which was causing test failures simply
because the output of the tests did not match expected output.