Fix: acq435st back from process to thread
Looks like the problems in: #1563 were related to garbage
collection not to a race condition.

This commit removes the process stuff and uses a defined
thread class to do the work.

I (Tom's idea) do not think there is a need to store the
device and or thread persistently in the class.  It may be
good enough for the thread to make a copy of the device.

Note:  It is sufficient for the init of the thread to copy its
argument, because the init is done sychronously before the thread
is started, so 'self' is still in scope during its execution