Fix:  acq2106_423_Nst devices to acq435 threads
In order to test the threading behavior under realistic conditions
we need at least 2 devices.  The acq435st was changed to use a
thread that copies the device object.

Issue: #1563 outlines some potential problems along these lines.

This commit changes to acq2106 to match the new acq435 code so they
can be tested on the SPARC network.