Fix: A commit message content can break the release tag jaon (#1610)
The release tag sent to github is a json which contains a list
of the new commits added to make the release. If this contains
certain characters it might break the json format which prevents
the release from being tagged. This first tries with the comments
and if that fails will just send the tag with a comment indicating
the changes could not be described in the tag info.