Build: Allow backslash in release tag
The release tag information includes the comments
from commits added since the previous release. If a commit
comment includes a backslash then this needs to be escaped or
it will produce a json error when sending the tag to github.
This should permit comments with backslashes to be included
in the release text.