Build: Remove dclunittest from stable
Currently the stable branch tends to get locked up doing the
mdsobjects/python/tests/dclUnitTest.py preventing the builds from
succeeding. This change disables that test for now. It is hoped that
the stable branch will be promoted to alpha in the near future which
will re-enable this test which does not have this problem in the alpha
branch.