Fix: TreeShr hook funs not working
The feature of enabling treeshr hooks as tdi funs was only working
when compiled without optimization. The code was declaring structures
within if else blocks and referencing those declarations later and
the optimizer was reusing the memory of those structures before
they were used. This commit provides the hook feature with
optimization enabled.