Fix: reinstate install of new _version.py's
With the release info now being generated by configure the risk of potential mismatch of version info in python and libraries with test builds has been eliminated. This is a change to address Andrea's request: "Please could you have a look on _version.py files? It seems to not be actually installed for a mismatch on destination path."