Feature: Reorganize MDSplus python module
The MDSplus python module was originally in a folder called
mdsobjects/python. This change moves the module to a directory
called python/MDSplus. This enables the ability to access the
module by using either the environment variable PYTHONPATH or
by installing an mdsplus.pth file in python's site-package
directory with the contents being the MDSplus python subdirectory.

The installers will in the future create a pth file for MDSplus.
If you want to add the MDSplus module using pip you will need to
use:

  pip install -e <mdsplus-module-dir>/setup.py

This will create an mdsplus.pth file pointing to the MDSplus
python directory in the site directory.

The main reason for this change is to avoid the problems that
occur when the installed MDSplus shared libraries are updated on
a system while users running their own python applications are
running an older version of the MDSplus python module which may
have become incompatible with the new libraries. With an mdsplus.pth
file which points to the installed MDSplus system's MDSplus python
module the module and libraries will remain compatible.