Feature: Reorganize python module
The directory structure for the MDSplus python module
is being changed from mdsplus/mdsobjects/python to
mdsplus/python/mdsplus. This enables the use of the module
by simply defining PYTHONPATH=<mdsplus-dir>/python.

The setup.py in the module has been changed to add an mdsplus.pth
file to the python site packages which directs python to reference
the module in the mdsplus/python/MDSplus directory similar to
what changing PYTHONPATH would do.