Fix: getOriginalPartName returns a str not String
getOriginalPartName used to return an MDSplus String object.
Several of the Mit Device models then called its getString method
to get a python string from it.

This was changed (when ?) to return a python string, which does
not have a getString method.

Since the python string is a more natural response, the device
code is now changed to assume it returns a python str.