Feature: Add node obeys usage when using .name syntax
In the past when adding a new node the code would look for a leading
period and would add a child node (usage structure) even if the usage
specified depicted a typical member node.

This change will create a child node only under the following conditions:

  1) The usage is "structure" or "subtree"
  2) The usage is "any" (default in most cases) and the node name is
     preceeded by a period.

Otherwise a member node will be created with the specified usage.