Feature:Use Compile in palce of Execute to return dimension in getSegment()
It may happen that getSegment() is used even in case the associate dimension is inconsistent. Using TdiExecute() for the returned dim in this case makes getSegment() fail or crash. The problem is solved by just compiling the returned dimension.  NOTE: dimension must be consisted when the whole signal ir retrieved, i.e. when xtreeshr is involved.