Feature: Add SANDBOX
This feature provides a SANDBOX setting which will disable the use
of the spawn function or the py() tdi function for security reasons.
The SANDBOX setting can be turned on using the MdsShr->MdsEnableSandbox()
function. Once enabled for the process it cannot be disabled. When
enabled the LibSpawn function will return and error and execution of the py()
function will also be disabled. You can now also add mdsip.hosts mapping rules
such as:

<match-string> | SANDBOX

and if the connection matches the string it will enable SANDBOX in the mdsip
process and use the nobody uid.