Fix: Restore menu updates for Windows
The latest Windows operating system updates seemed to deprecate
the use of a obsolete Windows message used to trigger the update
of the Start Menu items after new shortcuts are added or deleted during
a package installation or removal. This fix changes the message
type sent after the MDSplus installation to ensure the menus are
updated.