Feature: Convert nids in funs to paths
When a node reference is hardcoded in side a tdi fun the node
is converted into a nid (node id reference) at compile time when
the fun is first referenced. This will often cause problems if
subsequently the current process opens other trees since the nid
compiled into the function may not resolve to the same node in
the new tree context.

This change will automatically convert any nids found in a tdi fun
to a tree path instance. When the function is executed, instead of
trying to reference the node by a nid, the function will now find
the node referenced using a tree path string. This should avoid
the potential conflicts when the current tree context is changed
after the initial loading of the tdi function.