Fix:ScopeUtilities.c did not report correctly segmented info
jScope always asks to evaluate a sequence ( _ROW = <row>; _COLUMN = <col>; _INDEX = <idx>; <expression>) that is presented as a function descriptor. In this case the expression was always considered not segmented, thus preventin all the jScope machinery for improving efficiency in waveform access. Not the expression is recursively evaluated in order to edetect whether it is based on segmented nodes.