Feature:provide efficient event based redisplay for segmented signals
When events are used to update a growing signal, now only the new part of the signal is quesried by jScope to the data provider. In addition, internal structures are updated every time an event is received in order to avoid running out of heap space when many new updateds are performed.