Gm improve jscope (#1804)
* Fix:2D visualization

The added setting of TDI var for column and row was missing a final semicolo and it broke 2D signal definition

* Fix:Error messages and default node

Recent changes broke the exception chain for error display. Now the exception chain has been fixed. Moreover a previously existing bug prevented dynami resampling when default node were defined. It has been fixed now

* Fix:ScopeUtilities.c did not report correctly segmented info

jScope always asks to evaluate a sequence ( _ROW = <row>; _COLUMN = <col>; _INDEX = <idx>; <expression>) that is presented as a function descriptor. In this case the expression was always considered not segmented, thus preventin all the jScope machinery for improving efficiency in waveform access. Not the expression is recursively evaluated in order to edetect whether it is based on segmented nodes.

* Fix:crash when events set and connection goes down

If a MDS event was defined and the client disconnected, upon the next event occurrence a null connection was used thus generating a core dump. Now the connection is checked before serving the event and if null, mdsip disconnects from that event.

* Feature:provide efficient event based redisplay for segmented signals

When events are used to update a growing signal, now only the new part of the signal is quesried by jScope to the data provider. In addition, internal structures are updated every time an event is received in order to avoid running out of heap space when many new updateds are performed.

* Fix:handling of expressions referring to segmented nodes

The path name of the segmented node referred to in a jScope expression is now retrieved and used to get dimensions

* Fix:wrong limits report by estimateNumSamples in ScopeUtilities.c

Recent changes introduced this bu: the segment limits were reported instead of the real ones.

* Feature:Ask for reconnection after data provider crash

Valid only when -s or -l options have been set. Otherwise the server must be redefined using the Network dialog

* Fix:memory leak

Signal instances were not deallocated by the GC as they appeared listed in a WaveData array.

* Fix:memory leak

* Fix:logaritmic scales