Gm marte2 streaming (#1839)
* Feature:improve MARTe2-MDSplus framework

* Feature:Add DW Setup for MARTe2 devices

* Fix:initial debugging of MARTE2_COMPONENT and MARTE2_NI6259

* Feature:add Setup Form for MARTE2_NI6259

* Fix merge

* Fix merge

* Feature:Add Oscilloscope mode to streaming. Add new MARTe2 devices. Allow D&G in Device fields

* Fix:moved to  java/jdevices/src/main/java

* Feature:New MARTe devices setup

* Fix:Add forgoten buttons

* Fix:handle comma instead of point in floating point text (italian locale)

* Feature:Added DEMOADC

* Feature:Added DEMOSTREAM device for DEMO streaming purpose

* Feature:add events to DEMOSTREAM

* Fix:2D visualization

The added setting of TDI var for column and row was missing a final semicolo and it broke 2D signal definition

* Fix:Error messages and default node

Recent changes broke the exception chain for error display. Now the exception chain has been fixed. Moreover a previously existing bug prevented dynami resampling when default node were defined. It has been fixed now

* Fix:ScopeUtilities.c did not report correctly segmented info

jScope always asks to evaluate a sequence ( _ROW = <row>; _COLUMN = <col>; _INDEX = <idx>; <expression>) that is presented as a function descriptor. In this case the expression was always considered not segmented, thus preventin all the jScope machinery for improving efficiency in waveform access. Not the expression is recursively evaluated in order to edetect whether it is based on segmented nodes.

* Fix:crash when events set and connection goes down

If a MDS event was defined and the client disconnected, upon the next event occurrence a null connection was used thus generating a core dump. Now the connection is checked before serving the event and if null, mdsip disconnects from that event.

* Feature:provide efficient event based redisplay for segmented signals

When events are used to update a growing signal, now only the new part of the signal is quesried by jScope to the data provider. In addition, internal structures are updated every time an event is received in order to avoid running out of heap space when many new updateds are performed.

* Fix:handling of expressions referring to segmented nodes

The path name of the segmented node referred to in a jScope expression is now retrieved and used to get dimensions

* Fix:wrong limits report by estimateNumSamples in ScopeUtilities.c

Recent changes introduced this bu: the segment limits were reported instead of the real ones.

* Feature:Ask for reconnection after data provider crash

Valid only when -s or -l options have been set. Otherwise the server must be redefined using the Network dialog

* Fix:memory leak

Signal instances were not deallocated by the GC as they appeared listed in a WaveData array.

* Fix:memory leak

* Fix:logaritmic scales

* Fix:lost optimization upon recent changes

A bunch of fixes in order to enable smart resampling in XTreeGetTimedRecord()

* Fix:remove forgotten debug message

* Fix:handle multiple requests of updates for signal regions

In addition adjacent regions are merged in the request to optimize traffic

* Feature:lower threshold for smart resampling in jScope to 5MSamples

* Fix:forgotten active debug flag

* Fix:avoid writing null sized arrays

Avoid writing null sized arrays in putSegmentResamples() and putSegmentMinMax() that may happen when writing chunks of data with less than RESOLUTION_FACTOR samples

* Feature:make use of MinMax

Let jScope define env variable MDSPLUS_DEFAULT_RESAMPLE_MODE=MinMax to enable MinMax resampling in xtreeshr. Moreover, lowered threshold to enable xtreeshr resampling to 50000 samples.

* Fix:avoid possible division by 0

* Fix:wrong computation of desired delta in resampling

Now it takes into account the actual time window in jScope zoom

* Feature:improved MARTe2 device organization

A bunch of fixes to achieve a worling starting point for development

* Feature:mode jServer to dmsobject java classes

* Feature:Added EXPR and Python MARTe2 devices

* Fix:fixed several minor issues

* Build: updated updatepkg script to use samfc as bootstrap and modern debian os

* Build: updated packaging for rfx devices

* Fix:avoid compiler warning for unused JNI function argument

* Fix:undo previous change due to different usage of commas in Italian locale. Now Italian locale is used no more.

* Fix:remove linker error in WIndoes for javamds