Fix: This should fix the /IF option in TCL DO/METHOD
The /IF option in DO/METHOD had a bug when calling TdiIntrinsic similar to
an earlier bug in the handling of the /ARG option. This should fix that problem.

This also will add to the output HELP DO/METHOD to show the syntax used for
optionally providing a list of arguments in the /ARG option.