Fix: Improve help output formating
The xml used to provide the help output in mdsdcl application
must include spaces in the front of blank lines used in the help
message otherwise the blank line is removed. The blank lines were
added intentionally to the help message to better format the
contents but most did not include leading blanks and were omitted
from the output.