Fix: TCL DO/METHOD /IF option problem and TCL HELP improvements (#1863)
* Fix: This should fix the /IF option in TCL DO/METHOD

The /IF option in DO/METHOD had a bug when calling TdiIntrinsic similar to
an earlier bug in the handling of the /ARG option. This should fix that problem.

This also will add to the output HELP DO/METHOD to show the syntax used for
optionally providing a list of arguments in the /ARG option.

* Fix: Improve help output formating

The xml used to provide the help output in mdsdcl application
must include spaces in the front of blank lines used in the help
message otherwise the blank line is removed. The blank lines were
added intentionally to the help message to better format the
contents but most did not include leading blanks and were omitted
from the output.