Fix: fix sign compare warning on lex generated c c
Using the new bootstrap and attempting to then build MDSplus on
Ubuntu-10 with deploy/build.sh  the build fails because the code
generated by lex in mdsdcl and tdishr generates a warning.

I am not sure why our builds are succeeding, my guess is the
bison/flex on UB18 is newer or older than what was used for the
kits.