Feature: Apply clang-format to all .c, .h, .cpp
This PR applies clang-format to all of the rest of the c and c++
source files in mdsplus.

For now it uses the default llvm formatting.
This breaks instances where both
  windows.h
and
  winsock2.h
are included.  (winsock2.h has to be first)
in theory a .clang-format file specifying the 'priority' of regexes
of the includes could prevent this.

I ran the default config, and checked for this case by hand.