Fix: old FORTRAN programs pass 1 for len
Old MDSplus documentation for the fortran interface stated that
when calling MdsValue and not interested in the return length
argument to pass a "1".

An ugly fix which may break non fortran callers? prints a warning,
and does not attempt to overwrite a 1 passed by reference as the
last argument to MdsValue.

Updated fortran test programs to reflect this.