Fix: cpp not receiving new event based streams
This change from Darren Garnier changes the event name to a char*
from a c++ string.  Seems to fix the problem.

Note:  Now that each stream is its own event, it is its own thread
and all of the logic checking on, saving and sending the name is no
longer needed.  Ths PR does not address this.