build: ub20 does not have python_numpy
The dependecies in linux.xml  Only distinguish on 'platform'.
UB20 has different dependencies than UB<20.  Added new section to
linux.xml for specifying DIST dependencies and logic to
debian_build_debs.py to find and use this.

Eventually this should be refactored so that the 'DIST' section is
inside of the 'PLATFORM' section.

the RHEL8 hack (its own XML) can probably leverage this same thing.