Fix: on some systems UIDs are > 16 bits
On systems using active directory uids can be constructed from
AD SIDs and may have bits in their high word set.  This PR
addresses https://github.com/MDSplus/mdsplus/issues/2375

If their are bits set in the high word of the UID then do not or in
the group.

When displaying in TCL, if the the low 16 bits of the owner do not
translate to a user, then try to translate all 32 bits to an owner