Add RHEL9 build, remove xinetd package dependency, bug fixes
Add rhel9.opts to use mdsplus/builder:rhel9 (which is based on Rocky Linux 9.1, as CentOS 9 does not exist)
Remove xinetd as a package requirement in linux.xml.
If /etc/xinet.d/ is present, it will still install the config files, but the packages will no longer depend on it. This is because we now have systemd files as an alternative, and RHEL9 has officialy dropped support for it, so this seems like a good time to cut ties.
When installing the `mdsplus-*-kernel` package on a system without xinetd or systemd, the post install script will throw errors trying to copy files. Instead, we wrap them in if statements to only copy the files if those directories exist.
Replace `.getiterator()` with `.iter()` to fix error when using `xml.ElementTree` with python3 in `redhat_build_rpms.py` and `alpine_build_apks.py`
This already seemed to be fixed in `debian_build_debs.py`
Removed seemingly duplicate include for `libdc1394_support*.so` in linux.xml