Fix: dir /full segfault (#2647)
* Fix: dir /full segfault #2646

This bug was introduced when alternate compression methods were added
commit 9829bf64dd7f3b9158f4d98172594fc523e4b746

That code allocated the NCI byte `spare2` to hold the index of the
compression method.  It assumed that if it was not filled in it would be
zero.  This turned out not to be the case for some nodes in some trees.

This PR adds a DBI that says wether or not to look at this byte.  The
** the default value of this DBI is ignore the compression method byte. **
To enable alternate compression methods, which have been seldom used
(never?) set the DBI DbiADVANCED_COMPRESSION to true.

- open the tree for edit
- call TreeSetDbi with DbiADVANCED_COMPRESSION set to 1
- write the tree

There will be TCL verbs to set and get this database attribute.

The database characteristics (DBI) have been initialized to zero since
https://github.com/MDSplus/mdsplus/blame/263cc8bce631b291f17ae2c3e09164c55a067b93/treeshr/TreeOpen.c#L659

Example that sets the flag:
```
int main()
{
  int status;
  int length;
  int one = 1;
  DBI_ITM cmp_itm[2];
  const char *tree = "test";
  const int shot = 42;

  cmp_itm[0].buffer_length=1;
  cmp_itm[0].code = DbiADVANCED_COMPRESSION;
  cmp_itm[0].pointer = (void *)&zero;
  cmp_itm[0].return_length_address = &length;
  cmp_itm[1].code = DbiEND_OF_LIST;

  status = TreeOpenEdit(tree, shot);
  if (!(status & 1)) {
    printf("open - status = %d\n", status);
    exit(1);
  }
  status = TreeSetDbi(cmp_itm);
  if (!(status & 1)) {
    printf("SetDBI - status = %d\n", status);
    exit(1);
  }
  status = TreeWriteTree("nb", 196500);
  if (!(status & 1)) {
    printf("Write - status = %d\n", status);
    exit(1);
  }
}
```

---------

Co-authored-by: Stephen Lane-Walsh <slwalsh@psfc.mit.edu>
Co-authored-by: Fernando Santoro <fsantoro@psfc.mit.edu>