Fix: putMany
There are two issues fixed, the first is that the `EXPORT` on `PutManyExecute` was missing, leading to the `LibKEYNOTFOU` error when calling the method from TDI
The second is in `putManyObj`, where we were passing `argsData->getDscs()` as `mdsdsc_t` pointers, when they were in fact `Data` pointers.
There is now a temporary vector that holds the results of `convertToDsc()` for each `Data`, and then calls `freeDsc` at the end of the function.

This fixes #2813 and unblocks the mdsthin putMany test