Fix: Handling of java options in configure script
* Fix: Handling of java options in configure script

The first fixed issue was that whenever providing --disable-java or --enable-java with any value, the conditional ENABLE_JAVA was set to TRUE.
This meant that even with --disable-java running the tests would also run the java tests even though the java classes were not built.

The second fixed issue was that a warning message about a missing JDK was always displayed even with --disable-java.