Fix: C++ TreeNode remove unnecessary Tree duplication (#2861)
* Fix:Avoid Tree duplication

The same Tree instance is used now among TreeNodes created from the same Tree. This in order to avoid the
proliferation of new contextes and consequently of new open sokets in distributed configuration

* Fix:avoid proliferation of open trees

When creating a new tree node from a tree, the tree reference is passed straight. To avoid multiple free() calls the tree is not freed at TreeNode destructor

* Fix:compile error

added forgotten }