Fix: errors with asan and tsan
Update `strtok` to `strtok_r` in `ReplaceAliasTrees` to prevent issues when multithreading. The internal `saveptr` was being clobbered, so this moves it into the function, and therefore safely in the thread.
Reorder the `io_disconnect` function for the thread backend. The cleanup of the last request/response messages were being interrupted by the `pthread_cancel`, which is fixed by moving the `pipe_close` calls above it, matching the windows implementation.