Fix: Make sure top node's parent state is on
Fixes: https://github.com/MDSplus/mdsplus/issues/2953

if the parent of a subtree is turned off, the parent_state of its
top node is set to off.

Subsequently if the subtree is opened by itself, the top node's
parent_state is off.  Since it has no parent in this scenario its
parent_state is its own state.

Changed _TreeTurnOn to set the parent_state to on if the node in
question is the top node.

Also changes SetNci flags, if it changes the state flag, then it also
calls SetParentState