Fix: fix brother of subtree nodes. (#2974)
* Fix: fix brother of subtree nodes.

The optimizer is deciding that it does not need to fill in the
brother nid of nodes that are subtree attachment points.
Making the variable that holds it volatile forces the optimizer to
do the right thing.

Fixes: https://github.com/MDSplus/mdsplus/issues/2972

* Fix: undo fix publish debug -  subtree-brother fix

PR 2973 enabled debug for the release build to work around the issue
that is fixed by this PR.  This commit undoes that.

* Update treeshr/TreeOpen.c